Dev vaibhav

+++ server/controller/subCategory.controller.ts
@@ -1,5 +1,5 @@
 import { NextFunction, Request, Response } from "express";
-import { SubcategoryService } from "../services";
+import { CategoryService, SubcategoryService } from "../services";
 import { SubCategoriesValidation } from "../validations";
 import { SearchSubCategoryDTO, CreateSubCategoryDTO, EditSubCategoryDTO } from "../dto";
 import { DuplicateRecord, NotExistHandler } from "../errorHandler";
@@ -8,13 +8,14 @@ import { removeFile, saveFile } from "../utils/helper";
 import { Category, SubCategory } from "../models";
 
 export default class SubCategoryController {
-	private subCategoryService = new SubcategoryService();
-	private subCategoryValidations = new SubCategoriesValidation();
+	private service = new SubcategoryService();
+	private categoryService = new CategoryService();
+	private validations = new SubCategoriesValidation();
 
 	public getAll = {
-		validation: this.subCategoryValidations.getAll,
+		validation: this.validations.getAll,
 		controller: async (req: Request, res: Response, next: NextFunction): Promise<void> => {
-			const data = await this.subCategoryService.getAll(new SearchSubCategoryDTO(req.query));
+			const data = await this.service.getAll(new SearchSubCategoryDTO(req.query));
 			res.api.create(data);
 		},
 	};
@@ -27,19 +28,35 @@ export default class SubCategoryController {
 				throw new DuplicateRecord("Category not found");
 			}
 
-			const data = await this.subCategoryService.getSubCategoryByCategory(categoryData.id);
+			const data = await this.service.getSubCategoryByCategory(categoryData.id);
+			res.api.create(data);
+		},
+	};
+
+	public findOne = {
+		controller: async (req: Request, res: Response, next: NextFunction): Promise<void> => {
+			const categoryId: string = req.params["id"] as string;
+			const categoryExist = await Category.findByPk(categoryId, { attributes: ["id", "name", "details", "img_url", "logo_url"] });
+			if (!categoryExist) {
+				throw new NotExistHandler("Sub category Not Found");
+			}
+			const data = await this.service.findOne({ category_id: categoryId });
 			res.api.create(data);
 		},
 	};
 
 	public create = {
-		validation: this.subCategoryValidations.create,
+		validation: this.validations.create,
 		controller: async (req: Request, res: Response, next: NextFunction): Promise<void> => {
 			const subcategoryData = new CreateSubCategoryDTO(req.body);
-			const subcategoryExist = await this.subCategoryService.findOne({ name: subcategoryData.name, is_deleted: false });
+			const subcategoryExist = await this.service.findOne({ name: subcategoryData.name, is_deleted: false });
 
 			if (subcategoryExist && subcategoryExist != null) {
 				throw new DuplicateRecord("Subcategory already exists");
+			}
+			const checkCategory = await this.categoryService.findOne({ id: subcategoryData.category_id });
+			if (checkCategory == null) {
+				throw new DuplicateRecord("Category not found");
 			}
 			const file: any = req.files;
 			if (file) {
@@ -52,13 +69,13 @@ export default class SubCategoryController {
 					subcategoryData.logo_url = profile.upload_path;
 				}
 			}
-			const data = await this.subCategoryService.create(subcategoryData);
+			const data = await this.service.create(subcategoryData);
 			res.api.create(data);
 		},
 	};
 
 	public edit = {
-		validation: this.subCategoryValidations.edit,
+		validation: this.validations.edit,
 		controller: async (req: Request, res: Response, next: NextFunction): Promise<void> => {
 			const subcategoryId: string = req.params["id"] as string;
 			const subCategoryExist = await SubCategory.findByPk(subcategoryId);
@@ -66,7 +83,7 @@ export default class SubCategoryController {
 				throw new NotExistHandler("Sub Category Not Found");
 			}
 			const subcategoryData = new EditSubCategoryDTO(req.body);
-			const subcategoryExist = await this.subCategoryService.findOne({
+			const subcategoryExist = await this.service.findOne({
 				id: { [Op.not]: subcategoryId },
 				name: subcategoryData.name,
 				is_deleted: false,
@@ -75,25 +92,25 @@ export default class SubCategoryController {
 			if (subcategoryExist && subcategoryExist != null) {
 				throw new DuplicateRecord("Subcategory already exists");
 			}
+			const checkCategory = await this.categoryService.findOne({ id: subcategoryData.category_id });
+			if (checkCategory == null) {
+				throw new DuplicateRecord("Category not found");
+			}
 			const file: any = req.files;
 			if (file) {
-				const oldImgData = await this.subCategoryService.findOne({ id: subcategoryId });
+				const oldImgData = await this.service.findOne({ id: subcategoryId });
 				if (file.img_url) {
-					// Remove old image
 					oldImgData?.img_url && (await removeFile(oldImgData.img_url));
-					// Upload new image
 					let profile: any = await saveFile(file.img_url, "subcategory");
 					subcategoryData.img_url = profile.upload_path;
 				}
 				if (file.logo_url) {
-					// Remove old image
 					oldImgData?.logo_url && (await removeFile(oldImgData.logo_url));
-					// Upload new image
 					let profile: any = await saveFile(file.logo_url, "subcategory");
 					subcategoryData.logo_url = profile.upload_path;
 				}
 			}
-			const data = await this.subCategoryService.edit(subcategoryId, subcategoryData);
+			const data = await this.service.edit(subcategoryId, subcategoryData);
 			res.api.create(data);
 		},
 	};
@@ -105,14 +122,14 @@ export default class SubCategoryController {
 			if (!subCategoryExist) {
 				throw new NotExistHandler("Sub Category Not Found");
 			}
-			await this.subCategoryService
+			await this.service
 				.delete(subcategoryId, req.authUser.id)
-				.then(async (data) => {
-					if (data.img_url) {
-						await removeFile(data.img_url);
+				.then(async () => {
+					if (subCategoryExist.img_url) {
+						await removeFile(subCategoryExist.img_url);
 					}
-					if (data.logo_url) {
-						await removeFile(data.logo_url);
+					if (subCategoryExist.logo_url) {
+						await removeFile(subCategoryExist.logo_url);
 					}
 					res.api.create({
 						message: `Subcategory deleted`,
